Finalize into_hir conversion
