+++ optd/src/dsl/analyzer/from_ast/converter.rs
@@ -1,6 +1,7 @@
 use crate::dsl::analyzer::errors::AnalyzerErrorKind;
 use crate::dsl::analyzer::hir::{Annotation, FunKind, Identifier};
-use crate::dsl::analyzer::type_checks::registry::{Type, TypeRegistry, create_function_type};
+use crate::dsl::analyzer::type_checks::converter::create_function_type;
+use crate::dsl::analyzer::type_checks::registry::{Type, TypeRegistry};
 use crate::dsl::analyzer::{
     context::Context,
     hir::{CoreData, TypedSpan, Udf, Value},
