+++ optd/src/dsl/analyzer/from_ast/expr.rs
@@ -8,7 +8,8 @@ use crate::dsl::analyzer::errors::AnalyzerErrorKind;
 use crate::dsl::analyzer::hir::{
     BinOp, CoreData, Expr, ExprKind, FunKind, Identifier, LetBinding, Literal, TypedSpan, UnaryOp,
 };
-use crate::dsl::analyzer::type_checks::registry::{Type, TypeKind, create_function_type};
+use crate::dsl::analyzer::type_checks::converter::create_function_type;
+use crate::dsl::analyzer::type_checks::registry::{Type, TypeKind};
 use crate::dsl::parser::ast::{
     self, BinOp as AstBinOp, Expr as AstExpr, Literal as AstLiteral, PostfixOp,
 };
